refactor: 커스텀 훅으로 state 관리, 코드 정리 #22
